+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 1.0.1
 -------------------------
+- [Feature] Add php 7.3 and 7.4 to pipeline (alexeevdv)
 - [Bug] SmscRu | #1 Wrong API endpoint in send method (alexeevdv)
 
 1.0.0