+++ CHANGELOG.md
@@ -4,12 +4,12 @@ Under development
 - [Feature] Extract provider contract to separate package (alexeevdv)
 - [Feature] Use provider interface instead of base class (alexeevdv)
 
-1.0.1
+1.0.1 (2020-06-01)
 -------------------------
 - [Feature] Add php 7.3 and 7.4 to pipeline (alexeevdv)
 - [Bug] SmscRu | #1 Wrong API endpoint in send method (alexeevdv)
 
-1.0.0
+1.0.0 (2018-03-01)
 -------------------------
 - [SmsRu](https://sms.ru/) provider added
 - [SmscRu](https://smsc.ru/) provider added